add mysql commands + start changing pg commands

+++ model/user.go
@@ -4,5 +4,5 @@ type User struct {
 	ID       int64  `db:"id"`
 	Name     string `db:"name"`
 	UserType string `db:"utype"`
-	Balance  int64  `db:"balance"`
+	Balance  int    `db:"balance"`
 }
